Fix some signed/unsigned warnings

+++ libduel.cpp
@@ -4417,7 +4417,7 @@ int32 scriptlib::duel_assume_reset(lua_State *L) {
 }
 int32 scriptlib::duel_get_card_from_fieldid(lua_State * L) {
 	check_param_count(L, 1);
-	int32 fieldid = lua_tointeger(L, 1);
+	uint32 fieldid = static_cast<uint32>(lua_tointeger(L, 1));
 	duel* pduel = interpreter::get_duel_info(L);
 	for(auto& pcard : pduel->cards) {
 		if(pcard->fieldid_r == fieldid) {
